requirements.txt
numpy==1.22.4
scipy==1.10.1
pandas==1.5.3
scikit-learn
xgboost
textblob
spacy
python-dotenv
psycopg2-binary